+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/OpenId4VpSdJwtProtocolTest.kt
@@ -53,7 +53,7 @@ class OpenId4VpSdJwtProtocolTest : FreeSpec({
     "Selective Disclosure with custom credential" {
         val requestedClaim = AtomicAttribute2023.CLAIM_GIVEN_NAME
         val authnRequest = verifierOid4vp.createAuthnRequest(
-            RequestOptions(
+            OpenIdRequestOptions(
                 setOf(
                     RequestOptionsCredential(AtomicAttribute2023, SD_JWT, setOf(requestedClaim))
                 )
@@ -80,7 +80,7 @@ class OpenId4VpSdJwtProtocolTest : FreeSpec({
             EuPidScheme.SdJwtAttributes.GIVEN_NAME_BIRTH, // "birth_given_name" instead of "given_name_birth"
         )
         val authnRequest = verifierOid4vp.createAuthnRequest(
-            RequestOptions(
+            OpenIdRequestOptions(
                 credentials = setOf(
                     RequestOptionsCredential(EuPidScheme, SD_JWT, requestedClaims)
                 )