+++ vck-rqes/src/commonTest/kotlin/at/asitplus/wallet/lib/KotestConfig.kt
@@ -10,5 +10,6 @@ class KotestConfig : AbstractProjectConfig() {
         Napier.takeLogarithm()
         Napier.base(DebugAntilog())
         initRqesModule()
+        at.asitplus.wallet.eupid.Initializer.initWithVCK()
     }
 }