+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
 Release 5.4.0:
-- tbd
+- Extend support for POTENTIAL UC5: Remote qualified electronic signatures
+  - Update data classes in `rqes-data-classes`
+  - See main classes `RqesOpenId4VpHolder` and `RqesOpenId4VpVerifier` in `vck-rqes`
+  - Still open: Combination with OpenID4VP Draft 23
+- Refactorings:
+  - Rename `RequestOptions` to `OpenIdRequestOptions`
 
 Release 5.3.0:
 - Implement [token-status-list-06](https://www.ietf.org/archive/id/draft-ietf-oauth-status-list-06.html), replacing implementation of Revocation List 2020: